removed EmployeesService service class and moved methods into EmployeeService class
